requirements.txt
Flask==0.12.2
PyYAML==3.12
argparse
requests==2.18.4
gunicorn==19.7.1
six==1.11.0
